Rename the main package from eu.cessda.eqb.harvester to eu.cessda.oaiharvester
The application is no longer part of EQB, so references to EQB should be removed.
