+++ .gitlab-ci.yml
@@ -17,7 +17,7 @@ jar-test:
   after_script:
     - echo -n 'Branch coverage ' && grep -oE '<tfoot>.*</tfoot>' target/site/jacoco/index.html | grep -oE '[0-9]{1,3}%' | tail -1
   only:
-    - master
+    - main
     - stable
 
 gesis-maven:
@@ -25,7 +25,7 @@ gesis-maven:
   script:
     - mvn clean deploy -DskipTests
   only:
-    - master
+    - main
     - stable
 
 gesis-docker:
@@ -33,5 +33,5 @@ gesis-docker:
   script:
     - mvn clean -DskipTests package docker:build docker:push -Pdocker-compose
   only:
-    - master
+    - main
     - stable
