+++ pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>eu.cessda.eqb</groupId>
+	<groupId>eu.cessda</groupId>
 	<artifactId>oaiharvester</artifactId>
 	<version>3.2.2-SNAPSHOT</version>
 	<name>CESSDA OAI-PMH Metadata Harvester</name>
