+++ src/test/java/org/oclc/oai/harvester2/verb/ListRecordsTests.java
@@ -21,7 +21,7 @@ package org.oclc.oai.harvester2.verb;
  */
 
 
-import eu.cessda.eqb.harvester.HttpClient;
+import eu.cessda.oaiharvester.HttpClient;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.w3c.dom.Element;
